Floyd's cycle finding algorithm
Floyd's cycle-finding algorithm is a pointer algorithm that uses only two pointers, moving through the sequence at different speeds. It states the usage of Linked List in this algorithm and its output. The purpose is to determine whether the linked list has a cycle or not. First, you keep two pointers of the head node